PokeAByte.Domain: Split Exceptions.cs and moved them to their own folder.
