+++ src/PokeAByte.Domain/Exceptions/MapperException.cs
@@ -1,0 +1,6 @@
+namespace PokeAByte.Domain;
+
+public class MapperException : VisibleException
+{
+    public MapperException(string message, Exception? innerException = null) : base(message, innerException) { }
+}
