+++ src/PokeAByte.Domain/Exceptions/MapperInitException.cs
@@ -1,0 +1,6 @@
+namespace PokeAByte.Domain;
+
+public class MapperInitException : VisibleException
+{
+    public MapperInitException(string message, Exception? innerException = null) : base(message, innerException) { }
+}
