+++ src/PokeAByte.Domain/Exceptions/PropertyProcessException.cs
@@ -1,0 +1,6 @@
+namespace PokeAByte.Domain;
+
+public class PropertyProcessException : Exception
+{
+    public PropertyProcessException(string message, Exception? innerException) : base(message, innerException) { }
+}
