+++ src/PokeAByte.Domain/Exceptions/VisibleException.cs
@@ -1,0 +1,8 @@
+namespace PokeAByte.Domain;
+
+public class VisibleException : Exception
+    {
+        public VisibleException(string message, Exception? innerException = null) : base(message, innerException)
+        {
+        }
+    }
